# openapi.yaml
openapi: 3.0.3
info:
  title: Untapped Energy Application
  description: >-
    An application for better support of the ongoing learning initiatives of Untapped Energy.
  version: "0.8.0"
servers:
  - url: http://localhost/api:5001
    description: Local host
paths:
...
  /auth:
    get:
      summary: Log a user in
      description: Log a user in.
      security:
        - AuthHTTPBasic: []
      responses:
        "200":
          description: >-
            User logged in.  The token in the response body is the bearer token that represents
            the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found.
        "504":
          $ref: "#/components/responses/CantConnectToDatabase"

    post:
      summary: Add a new user.
      description: Add a new user to the system.
      security:
        - AuthHTTPBasic: []
      requestBody:
        description: Details of the user to be added.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInfo"
      responses:
        "201":
          description: >-
            User successfully added.  The token in the response body is the bearer token that
            represents the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "403":
          description: User already exists.
        "406":
          $ref: "#/components/responses/InvalidData"
        "504":
          $ref: "#/components/responses/CantConnectToDatabase"
...
  /user:
    get:
      summary: Return a specific user.
      description: Returns information about the user identified by the bearer token.
      security:
        - AuthHTTPBearer: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "504":
          $ref: "#/components/responses/CantConnectToDatabase"

    patch:
      summary: Update a specific user.
      description: Update information about the user identified by the bearer token.
      requestBody:
        description: Updated details of the user.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserData"
      security:
        - AuthHTTPBearer: []
      responses:
        "200":
          description: User successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "406":
          $ref: "#/components/responses/InvalidData"
        "504":
          $ref: "#/components/responses/CantConnectToDatabase"

    delete:
      summary: Delete a specific user.
      description: >-
        Removes the user identified by the bearer token from the system (as well as any related
        data).
      security:
        - AuthHTTPBearer: []
      responses:
        "200":
          description: User successfully deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "504":
          $ref: "#/components/responses/CantConnectToDatabase"
...
components:
  securitySchemes:
    AuthHTTPBasic:
      type: http
      scheme: basic
    AuthHTTPBearer:
      type: http
      scheme: bearer
  responses:
    CantConnectToDatabase:
      description: Can"t connect to database.
    InvalidData:
      description: Invalid data.
    Unauthorized:
      description: Not authorized.
  schemas:
    RegistrationInfo:
      type: object
      properties:
        name:
          type: string
        isInstructor:
          type: boolean
      required:
        - name
        - isInstructor
    Credential:
      type: object
      properties:
        # Naming conventions in this section comply with RFC 6749 (The OAuth 2.0 Authorization
        # Framework) which may be used later -- see
        # https://datatracker.ietf.org/doc/html/rfc6749#section-5.1

        access_token:
          type: string
        token_type:
          type: string
          enum:
            - "Bearer"
        user_data:
          $ref: "#/components/schemas/UserData"
      required:
        - access_token
        - token_type
        - user_data
    UserData:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        learnerData:
          $ref: "#/components/schemas/LearnerData"
          nullable: true
        instructorData:
          $ref: "#/components/schemas/InstructorData"
          nullable: true
      required:
        - name
        - email
        - learnerData
    LearnerData:
      type: array
      items:
        $ref: "#/components/schemas/Course"
    InstructorData:
      type: array
      items:
        $ref: "#/components/schemas/Course"
    Course:
      type: object
      properties:
        UID:
          type: string
        title:
          type: string
        description:
          type: string
        instructor:
          type: string
        components:
          type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/Module"
              - $ref: "#/components/schemas/Assessment"
        currentComponent:
          type: integer
        creationTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
      required:
        - UID
        - title
        - description
        - components
        - instructor
        - creationTime
        - updateTime
    Module:
      type: object
      properties:
        UID:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - "Audio"
            - "Video"
            - "Markdown"
        chapters:
          type: array
          items:
            oneOf:
              - type: string
              - type: number
        url:
          type: string
        urlAuthentication:
          type: object
          properties:
            scheme:
              type: string
            parameters:
              type: string
          required:
            - scheme
            - parameters
        creationTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
      required:
        - UID
        - title
        - description
        - type
        - url
        - creationTime
        - updateTime
    Assessment:
      type: object
      properties:
        UID:
          type: string
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        currentQuestion:
          type: integer
        creationTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
      required:
        - UID
        - title
        - questions
        - creationTime
        - updateTime
    Question:
      type: object
      properties:
        UID:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 26
        answer:
          type: integer
        correctChoice:
          type: integer
        explanation:
          type: string
        creationTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
      required:
        - UID
        - question
        - options
        - correctChoice
        - explanation
        - creationTime
        - updateTime